fix(typecheck): move role-menu builders to features/, drop type:ignore
mypy follows imports from src/, and the new src/webui/routes/rolemenus.py
imported from extensions.reactionroles._common — pulling the mixin-pattern
attr-defined holes into the typecheck scope (they're invisible elsewhere
because mypy doesn't follow into extensions otherwise).

Move build_embed / build_components / MAX_ENTRIES to
features/reactionroles/builders.py so both the slash-command extension and
the WebUI route import from a typed, src-adjacent module. Also drop the
unused # type: ignore[misc] on enabled_servers_int.
