+++ features/reactionroles/__init__.py
@@ -1,8 +1,18 @@
+from features.reactionroles.builders import (
+    BUTTON_PREFIX,
+    MAX_ENTRIES,
+    build_components,
+    build_embed,
+)
 from features.reactionroles.models import RoleMenu, RoleMenuEntry
 from features.reactionroles.repository import ReactionRolesRepository
 
 __all__ = [
+    "BUTTON_PREFIX",
+    "MAX_ENTRIES",
     "ReactionRolesRepository",
     "RoleMenu",
     "RoleMenuEntry",
+    "build_components",
+    "build_embed",
 ]
